app/requirements.txt
fastapi
uvicorn[standard]
whisperx
torch
torchaudio
python-dotenv